Consolidate context files
