--- src/contexts/Balance/Context.ts
@@ -1,5 +1,0 @@
-import { createContext } from 'react';
-
-import { ContextValues } from './types';
-
-export const BalanceContext = createContext<ContextValues>({});
