--- src/contexts/Balance/index.ts
@@ -1,2 +1,0 @@
-export { BalanceContext } from './Context';
-export { BalanceProvider } from './Provider';
