--- src/contexts/Balance/types.ts
@@ -1,3 +1,0 @@
-export interface ContextValues {
-  wpoktBalance?: any;
-}
