--- src/contexts/Web3/index.ts
@@ -1,2 +1,0 @@
-export { Web3Context } from './Context';
-export { Web3Provider } from './Provider';
